Introduce the Partial Prerendering(still experimental)
a new rendering model that allows you to combine the benefits of static and dynamic rendering in the same route.
As long as you're using Suspense to wrap the dynamic parts of your route, Next.js will know which parts of your route are static and which are dynamic.(Wrapping a component in Suspense doesn't make the component itself dynamic, but rather Suspense is used as a boundary between your static and dynamic code.
)