Chapter 11 Adding Search(including Debouncing) and Pagination
Learn how to use the Next.js APIs: useSearchParams, usePathname, and useRouter.

Implement search and pagination using URL search params.

NOTE:
1. When to use the useSearchParams() hook vs. the searchParams prop?
<Search> is a Client Component, so you used the useSearchParams() hook to access the params from the client.
<Table> is a Server Component that fetches its own data, so you can pass the searchParams prop from the page to the component.
As a general rule, if you want to read the params from the client, use the useSearchParams() hook as this avoids having to go back to the server.

Summary:

1. handled search and pagination with URL search parameters instead of client state.
2. fetched data on the server.
3. using the useRouter router hook for smoother, client-side transitions.