Chapter 13 Handling Errors
There are a few things you'll notice about the code above:

"use client" - error.tsx needs to be a Client Component.
It accepts two props:
error: This object is an instance of JavaScript's native Error object.
reset: This is a function to reset the error boundary. When executed, the function will try to re-render the route segment.

TO BE MORE SPECIFIC:
notFound will take precedence over error.tsx, so you can reach out for it when you want to handle more specific errors